--- uts/src/test/kotlin/io/ably/lib/types/Utils.kt
@@ -1,3 +1,0 @@
-package io.ably.lib.types
-
-fun ConnectionDetails(init: ConnectionDetails.() -> Unit) = ConnectionDetails().apply(init)
