+++ uts/src/test/kotlin/io/ably/lib/uts/infra/unit/ClientFactories.kt
@@ -1,4 +1,4 @@
-package io.ably.lib.uts.infra
+package io.ably.lib.uts.infra.unit
 
 import io.ably.lib.debug.DebugOptions
 import io.ably.lib.realtime.AblyRealtime
