+++ uts/src/test/kotlin/io/ably/lib/uts/infra/unit/FakeClock.kt
@@ -1,4 +1,4 @@
-package io.ably.lib.uts.infra
+package io.ably.lib.uts.infra.unit
 
 import io.ably.lib.util.Clock
 import io.ably.lib.util.AblyTimer
