+++ uts/src/test/kotlin/io/ably/lib/uts/infra/unit/MockHttpClient.kt
@@ -1,4 +1,4 @@
-package io.ably.lib.uts.infra
+package io.ably.lib.uts.infra.unit
 
 import io.ably.lib.debug.DebugOptions
 import io.ably.lib.network.HttpEngine
