+++ uts/src/test/kotlin/io/ably/lib/uts/infra/unit/PendingRequest.kt
@@ -1,4 +1,4 @@
-package io.ably.lib.uts.infra
+package io.ably.lib.uts.infra.unit
 
 import java.net.URL
 import kotlin.time.Duration
